+++ CHANGES.rst
@@ -2,6 +2,61 @@ Changelog
 =========
 
 
+1.1.0 (2023-06-30)
+~~~~~~~~~~~~~~~~~~
+
+New features
+------------
+
++ `#113`_, `#123`_: Add module :mod:`icat.ingest`.
+
++ `#124`_: Add an optional keyword argument `keepInstRel` to
+  :meth:`icat.entity.Entity.truncateRelations`.
+
+Bug fixes and minor changes
+---------------------------
+
++ `#126`_, `#127`_: Update outdated documentation.
+
++ `#112`_, `#118`_: Extend icatdata XSD adding extra attributes to
+  reference objects.
+
++ `#111`_, `#121`_: Change the type of
+  :attr:`icat.client.Client.Register` to
+  :class:`weakref.WeakValueDictionary`, fixing a memory leak.
+
++ `#119`_, `#120`_: Remove `_config` attribute from
+  :class:`icat.config.Configuration`.
+
++ `#115`_, `#116`_: Fix the test suite to work if either PyYAML or
+  lxml is not available.
+
++ `#128`_: Return an empty list from
+  :func:`icat.dump_queries.getDataPublicationQueries` when talking to
+  an ICAT server older than 5.0.
+
++ `#117`_: Fixed deprecation warnings from upcoming Python 3.12.
+
++ `#129`_: Review the build of the documentation at Read the Docs.
+
+.. _#111: https://github.com/icatproject/python-icat/issues/111
+.. _#112: https://github.com/icatproject/python-icat/issues/112
+.. _#113: https://github.com/icatproject/python-icat/issues/113
+.. _#115: https://github.com/icatproject/python-icat/issues/115
+.. _#116: https://github.com/icatproject/python-icat/pull/116
+.. _#117: https://github.com/icatproject/python-icat/pull/117
+.. _#118: https://github.com/icatproject/python-icat/pull/118
+.. _#119: https://github.com/icatproject/python-icat/issues/119
+.. _#120: https://github.com/icatproject/python-icat/pull/120
+.. _#121: https://github.com/icatproject/python-icat/pull/121
+.. _#123: https://github.com/icatproject/python-icat/pull/123
+.. _#124: https://github.com/icatproject/python-icat/pull/124
+.. _#126: https://github.com/icatproject/python-icat/issues/126
+.. _#127: https://github.com/icatproject/python-icat/pull/127
+.. _#128: https://github.com/icatproject/python-icat/pull/128
+.. _#129: https://github.com/icatproject/python-icat/pull/129
+
+
 1.0.0 (2022-12-21)
 ~~~~~~~~~~~~~~~~~~